Move non-route related functions out of __init__.py
